rename to detailModal.js
rename to detailModal.js
