+++ src/index.js
@@ -1,4 +1,4 @@
 import './js/main';
 import './sass/main.scss';
 import './js/keyword_searcher';
-import './js/services/detaileModal';
+import './js/services/detailModal';
