+++ src/js/services/detailModal.js
